add: clean index on file remove
